Enable/Disable RStudio Package Manager

Functions to enable or disable RSPM repos as well as the integration of install_sysreqs into install.packages and update.packages. When enabled, binary packages are installed from RSPM if available, and system requirements are transparently resolved and installed without root privileges.

Usage

enable()

disable()

Value

No return value, called for side effects.

Details

To enable rspm permanently, include the following into your .Rprofile:

suppressMessages(rspm::enable())

See also

Examples

if (FALSE) { # \dontrun{
# install 'units' and all its dependencies from the system repos
rspm::enable()
install.packages("units")

# install packages again from CRAN
rspm::disable()
install.packages("errors")
} # }